integrations/engram-overrides/requirements.txt
fastapi[all]>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
rdflib>=7.0.0
sqlalchemy>=2.0.0
sqlmodel>=0.0.16
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
owlready2>=0.45
jsonschema>=4.21.0
pyDatalog>=0.17.4
aiohttp>=3.13.3
httpx>=0.27.0
networkx>=3.2.0
PyJWT>=2.12.0
slowapi>=0.1.9
structlog>=24.1.0
prometheus-fastapi-instrumentator>=7.0.0
prometheus_client>=0.20.0
redis>=5.0.0
scikit-learn>=1.4.0
joblib>=1.3.0
textual>=0.51.0
rich>=13.7.0
pyyaml>=6.0.1
pyswip>=0.2.10
ccxt>=4.2.10
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
requests>=2.31.0
cryptography>=42.0.0
tenacity>=8.3.0
alembic>=1.13.0
prance>=23.6.21.0
openapi-spec-validator>=0.7.1
strawberry-graphql>=0.219.2
docker>=7.0.0
mcp>=0.1.0
# phi-3-client — not on PyPI; upstream placeholder; omitted for reproducible Docker builds
